refactor: alterando a estrutura de pastas dos jogos
